Make signature be based on sha512 of file content
This way we don't have to store the entire file in memory, and we
don't have to do many passes over the entire content when
e.g. validating against multiple keys.

In the install case we now keep the content fd open and do the copy
from the fd instead of the loaded content.

In general, this allows us to scale to larger files without unnecessary
memory use and load.