install: Don't create unused directories
Unless a (valid) file is copied into place, don't create the parent
directories of it. Otherwise you can use this to create arbitrarily
named directories, which may be a security problem.

Only if some file is valid with a particular path then is it safe to
create the path.

Also add a test that verifies this.